improve(Achats): stylisation du formulaire (#6760)
