+++ 2024-frontend/src/data/documentation.json
@@ -1,8 +1,10 @@
 {
   "accueil": "https://ma-cantine.crisp.help/fr/",
+  "ajouterAchat": "https://ma-cantine.crisp.help/fr/article/ajouter-manuellement-un-produit-via-loutil-de-suivi-des-achats-egalim-zdbhv7/",
   "bienDemarrer": "https://ma-cantine.crisp.help/fr/category/bien-demarrer-1fn9r2h/",
   "calculerNombreCouverts": "https://ma-cantine.crisp.help/fr/article/bien-calculer-son-nombre-de-couverts-1w58ozp/?bust=1766138160563",
   "correctionCampaign": "https://ma-cantine.crisp.help/fr/article/la-campagne-de-correction-pour-ma-teledeclaration-faq-1sk8ipw/",
+  "critèresQualiteDurabiliteProduits": "https://ma-cantine.crisp.help/fr/category/qualite-et-durabilite-des-produits-e0vos3/",
   "diversificationSourcesProteines": "https://ma-cantine.crisp.help/fr/article/comprendre-mes-obligations-diversification-des-sources-de-proteines-et-menu-vegetarien-1om8is0/",
   "groupesRestaurantsSatellites": "https://ma-cantine.crisp.help/fr/article/les-groupes-de-restaurants-satellites-definition-et-gestion-82qt51/?bust=1767947700084",
   "importBilanDetaille": "https://ma-cantine.crisp.help/fr/article/teledeclaration-detaillee-tableau-recapitulatif-des-champs-a-renseigner-h7oaqf/",
